requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
dateparser==0.7.6
defusedxml==0.7.0rc1
dj-static==0.0.6
Django==3.1
django-admin-tools==0.9.0
django-filter==2.3.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
install==1.3.3
oauthlib==3.1.0
Pillow==7.2.0
prompt-toolkit==2.0.9
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-core==3.3.3
SQLAlchemy==1.3.19
sqlparse==0.3.1
static3==0.7.0
tornado==5.1.1
tzlocal==2.1
urllib3==1.25.10
wcwidth==0.2.5
